+++ changelog.md
@@ -23,6 +23,7 @@
 - ImageInterface::setBlendingColor() was renamed to ImageInterface::setBackgroundColor()
 - ImageInterface::blendingColor() was renamed to ImageInterface::backgroundColor()
 - ImageInterface::pad() was renamed to ImageInterface::containDown()
+- ImageInterface::encode() has a different signature
 - Config::class option blendingColor was renamed to backgroundColor
 - BlendTransparencyModifer::class was renamed to FillTransparentAreasModifier::class
 - Changed default value for `background` to `null` in ImageInterface::rotate()